+++ CHANGELOG.rst
@@ -1,6 +1,13 @@
 Changelog
 #########
 
+1.0.1
+*****
+
+- removed dependency on the ctype extension
+- only ASCII letters are allowed in ``X-HTTP-Method-Override``
+
+
 1.0.0
 *****